Fix product refresh after user favorites (from collections page)
